vercel-api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.7.1
pydantic-settings==2.2.1
python-multipart==0.0.7
sqlalchemy==2.0.32
sqlmodel==0.0.16
aiosqlite==0.20.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
httpx==0.27.0
python-dotenv==1.0.1
pytest==7.4.0
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==24.2.0
flake8==7.0.0
isort==5.13.2
gunicorn==21.2.0
pydantic[email]==2.7.1
pydantic[dotenv]==2.7.1
psycopg2-binary==2.9.9
asyncpg==0.29.0
fpdf2==2.7.7
markdown2==2.4.10
alembic==1.13.1
passlib==1.7.4
bcrypt==4.0.1
PyJWT==2.8.0
google-generativeai==0.3.1
python-docx==1.1.2
pdfminer.six